Major Overhaul
Fixed the modular page to be responsive down to iPhone size.

Changed the blog post header section to be more different from Antimatter.

Replaced the Grav brand font with the Grav symbol icon.

Changed color scheme and references to the changed icon font set in Materializer.

Centered text in buttons with and without icons.